+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### 2026-09-17
+
+#### Added
+- Added `<leader>ot` to toggle automatic comment continuation per buffer; it remains disabled by default.
+
 ### 2026-09-16
 
 #### Added